+++ src/Sensors/IMU/IMU.cpp
@@ -67,27 +67,27 @@ namespace mmfs
 
     void IMU::packData()
     {
-        _Float16 accX = _Float16(measuredAcc.x());
-        _Float16 accY = _Float16(measuredAcc.y());
-        _Float16 accZ = _Float16(measuredAcc.z());
-        _Float16 quatX = _Float16(orientationGlobal.x());
-        _Float16 quatY = _Float16(orientationGlobal.y());
-        _Float16 quatZ = _Float16(orientationGlobal.z());
-        _Float16 quatW = _Float16(orientationGlobal.w());
+        float accX = float(measuredAcc.x());
+        float accY = float(measuredAcc.y());
+        float accZ = float(measuredAcc.z());
+        float quatX = float(orientationGlobal.x());
+        float quatY = float(orientationGlobal.y());
+        float quatZ = float(orientationGlobal.z());
+        float quatW = float(orientationGlobal.w());
 
         int offset = 0;
-        memcpy(packedData + offset, &accX, sizeof(_Float16));
-        offset += sizeof(_Float16);
-        memcpy(packedData + offset, &accY, sizeof(_Float16));
-        offset += sizeof(_Float16);
-        memcpy(packedData + offset, &accZ, sizeof(_Float16));
-        offset += sizeof(_Float16);
-        memcpy(packedData + offset, &quatX, sizeof(_Float16));
-        offset += sizeof(_Float16);
-        memcpy(packedData + offset, &quatY, sizeof(_Float16));
-        offset += sizeof(_Float16);
-        memcpy(packedData + offset, &quatZ, sizeof(_Float16));
-        offset += sizeof(_Float16);
-        memcpy(packedData + offset, &quatW, sizeof(_Float16));
+        memcpy(packedData + offset, &accX, sizeof(float));
+        offset += sizeof(float);
+        memcpy(packedData + offset, &accY, sizeof(float));
+        offset += sizeof(float);
+        memcpy(packedData + offset, &accZ, sizeof(float));
+        offset += sizeof(float);
+        memcpy(packedData + offset, &quatX, sizeof(float));
+        offset += sizeof(float);
+        memcpy(packedData + offset, &quatY, sizeof(float));
+        offset += sizeof(float);
+        memcpy(packedData + offset, &quatZ, sizeof(float));
+        offset += sizeof(float);
+        memcpy(packedData + offset, &quatW, sizeof(float));
     }
 }
